chore: rewrite widget data actions
